02-azure-functions/functions/requirements.txt
# Azure Functions Python Requirements

azure-functions>=1.18.0
azure-storage-blob>=12.19.0
azure-storage-queue>=12.9.0
applicationinsights>=0.11.10
opencensus-ext-azure>=1.1.13
